Style improvement

+++ src/io/shapez/Board.java
@@ -305,14 +305,14 @@ public class Board extends JPanel implements ActionListener, MouseWheelListener,
         int offX = cX % GlobalConfig.mapChunkSize < 0 ? cX % GlobalConfig.mapChunkSize + 16 : cX % GlobalConfig.mapChunkSize;
         int offY = cY % GlobalConfig.mapChunkSize < 0 ? cY % GlobalConfig.mapChunkSize + 16 : cY % GlobalConfig.mapChunkSize;
 
-
-
         if (currentChunk.contents[offX][offY] == null) {
             currentChunk.contents[offX][offY] = new Entity(item, tileTexture, rotation, cX, cY);
             usedChunks.add(currentChunk);
         }
-        int result = 0;
+
+        byte result = 0;
         result = checkSpecialProperties(currentChunk, offX,offY, item);
+
         if(result == 1){
             System.out.println(
                     "Tile of type " + item.toString() + "has invalid placement at " + offX + " " + offY + "\n" +
